backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.5.0,<2.10.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
dotenv>=0.9.9
python-multipart>=0.0.6  # Required for FastAPI Form data handling

# Development and testing
pytest>=8.4.1
pytest-asyncio>=1.1.0
httpx>=0.27.0

# ASW Lambda
mangum==0.19.0

# Crypto
cryptography


# database module dependencies
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
aiosqlite>=0.20.0
greenlet

# auth module dependencies
python-jose[cryptography]>=3.3.0

# aihub module dependencies
openai>=1.40.0,<1.60.0
sse-starlette>=1.6.0

# payment module dependencies
stripe>=12.0.0

# PDF extraction with coordinates
PyMuPDF>=1.24.0

# Word document support
python-docx>=1.1.0

# Google OAuth
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
reportlab>=4.0.0
email-validator>=2.0.0
